+++ gandi/cli/tests/commands/test_domain.py
@@ -44,9 +44,51 @@ updated     : 2014-09-21 03:10:07
 """)
         self.assertEqual(result.exit_code, 0)
 
-    def test_create(self):
+    def test_create_no_option_no_argument(self):
+        args = ['--duration', 1,
+                '--owner', 'OWNER1-GANDI',
+                '--admin', 'ADMIN1-GANDI',
+                '--tech', 'TECH1-GANDI',
+                '--bill', 'BILL1-GANDI',
+                ]
+        result = self.invoke_with_exceptions(domain.create, args,
+                                             input='idontlike.website\n')
+
+        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
+        self.assertEqual(output, """\
+Name of the domain: idontlike.website
+Creating your domain.
+\rProgress: [###] 100.00%  00:00:00  \n\
+Your domain idontlike.website has been created.""")
+
+        self.assertEqual(result.exit_code, 0)
+#         self.assertEqual(result.output.strip(), """\
+# Name of the domain.:\
+# """)
+
+    def test_create_option(self):
         result = self.invoke_with_exceptions(domain.create,
                                              ['--domain', 'idontlike.website',
+                                              '--duration', 1,
+                                              '--owner', 'OWNER1-GANDI',
+                                              '--admin', 'ADMIN1-GANDI',
+                                              '--tech', 'TECH1-GANDI',
+                                              '--bill', 'BILL1-GANDI',
+                                              ])
+
+        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
+        self.assertEqual(output, """\
+/!\ --domain option is deprecated and will be removed upon next release.
+You should use 'gandi domain create idontlike.website' instead.
+Creating your domain.
+\rProgress: [###] 100.00%  00:00:00  \n\
+Your domain idontlike.website has been created.""")
+
+        self.assertEqual(result.exit_code, 0)
+
+    def test_create_argument(self):
+        result = self.invoke_with_exceptions(domain.create,
+                                             ['idontlike.website',
                                               '--duration', 1,
                                               '--owner', 'OWNER1-GANDI',
                                               '--admin', 'ADMIN1-GANDI',
@@ -59,6 +101,71 @@ updated     : 2014-09-21 03:10:07
 Creating your domain.
 \rProgress: [###] 100.00%  00:00:00  \n\
 Your domain idontlike.website has been created.""")
+
+        self.assertEqual(result.exit_code, 0)
+
+    def test_create_argument_and_option_different(self):
+        result = self.invoke_with_exceptions(domain.create,
+                                             ['idontlike.website',
+                                              '--domain', 'idontlike.bike',
+                                              '--duration', 1,
+                                              '--owner', 'OWNER1-GANDI',
+                                              '--admin', 'ADMIN1-GANDI',
+                                              '--tech', 'TECH1-GANDI',
+                                              '--bill', 'BILL1-GANDI',
+                                              ])
+
+        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
+        self.assertEqual(output, """\
+/!\ --domain option is deprecated and will be removed upon next release.
+You should use 'gandi domain create idontlike.bike' instead.
+/!\ You specified both an option and an argument which are different, \
+please choose only one between: idontlike.bike and idontlike.website.""")
+
+        self.assertEqual(result.exit_code, 0)
+
+    def test_create_argument_and_option_equal(self):
+        result = self.invoke_with_exceptions(domain.create,
+                                             ['idontlike.website',
+                                              '--domain', 'idontlike.website',
+                                              '--duration', 1,
+                                              '--owner', 'OWNER1-GANDI',
+                                              '--admin', 'ADMIN1-GANDI',
+                                              '--tech', 'TECH1-GANDI',
+                                              '--bill', 'BILL1-GANDI',
+                                              ])
+
+        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
+        self.assertEqual(output, """\
+/!\ --domain option is deprecated and will be removed upon next release.
+You should use 'gandi domain create idontlike.website' instead.
+Creating your domain.
+\rProgress: [###] 100.00%  00:00:00  \n\
+Your domain idontlike.website has been created.""")
+
+        self.assertEqual(result.exit_code, 0)
+
+    def test_create_background_argument(self):
+        args = ['roflozor.com', '--background']
+        result = self.invoke_with_exceptions(domain.create, args)
+
+        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
+        self.assertEqual(output, """\
+Duration [1]: \n\
+{'id': 400, 'step': 'WAIT'}""")
+
+        self.assertEqual(result.exit_code, 0)
+
+    def test_create_background_option(self):
+        args = ['--domain', 'roflozor.com', '--background']
+        result = self.invoke_with_exceptions(domain.create, args)
+
+        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
+        self.assertEqual(output, """\
+Duration [1]: \n\
+/!\ --domain option is deprecated and will be removed upon next release.
+You should use 'gandi domain create roflozor.com' instead.
+{'id': 400, 'step': 'WAIT'}""")
 
         self.assertEqual(result.exit_code, 0)
 
@@ -88,6 +195,7 @@ Duration [1]: \n\
         self.assertEqual(result.exit_code, 0)
 
     def test_available_with_exception(self):
+
         self.assertRaises(DomainNotAvailable,
                           self.invoke_with_exceptions, domain.create,
                           ['--domain', 'unavailable1.website',
@@ -98,13 +206,13 @@ Duration [1]: \n\
                            '--bill', 'BILL1-GANDI',
                            ])
 
-    def test_create_background_ok(self):
-        args = ['--domain', 'roflozor.com', '--background']
-        result = self.invoke_with_exceptions(domain.create, args)
-
-        output = re.sub(r'\[#+\]', '[###]', result.output.strip())
-        self.assertEqual(output, """\
-Duration [1]: \n\
-{'id': 400, 'step': 'WAIT'}""")
-
-        self.assertEqual(result.exit_code, 0)
+    def test_available_with_exception_argument(self):
+        self.assertRaises(DomainNotAvailable,
+                          self.invoke_with_exceptions, domain.create,
+                          ['unavailable1.website',
+                           '--duration', 1,
+                           '--owner', 'OWNER1-GANDI',
+                           '--admin', 'ADMIN1-GANDI',
+                           '--tech', 'TECH1-GANDI',
+                           '--bill', 'BILL1-GANDI',
+                           ])